Implement extended ui.locators to Fix #3606
