+++ robottelo/ui/locators/base.py
@@ -1,846 +1,9 @@
 # -*- encoding: utf-8 -*-
 """Implements different locators for UI"""
 
-import collections
-import logging
-
 from selenium.webdriver.common.by import By
+from .model import LocatorDict
 
-
-LOGGER = logging.getLogger(__name__)
-
-
-class LocatorDict(collections.Mapping):
-    """This class will log every time an item is selected
-
-    The constructor accepts a dictionary or keyword arguments like the
-    built-in ``dict``::
-
-        >>> dict({'a': 'b'})
-        {'a': 'b'}
-        >>> dict(a='b')
-        {'a': 'b'}
-
-    """
-    def __init__(self, *args, **kwargs):
-        self.store = dict(*args, **kwargs)
-
-    def __getitem__(self, key):
-        item = self.store[key]
-        LOGGER.debug(
-            'Accessing locator "%s" by %s: "%s"', key, item[0], item[1]
-        )
-        return item
-
-    def __len__(self):
-        return len(self.store)
-
-    def __iter__(self):
-        return iter(self.store)
-
-
-menu_locators = LocatorDict({
-    # Menus
-    # Navbar
-    "navbar.spinner": (By.XPATH, ("//div[@id='turbolinks-progress']")),
-
-    # Monitor Menu
-    "menu.monitor": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='monitor_menu']")),
-    "menu.dashboard": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_dashboard']")),
-    "menu.reports": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_reports']")),
-    "menu.facts": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_fact_values']")),
-    "menu.statistics": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_statistics']")),
-    "menu.trends": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_trends']")),
-    "menu.audits": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_audits']")),
-
-    "menu.jobs": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_jobs']")),
-
-    # Content Menu
-    "menu.content": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='content_menu']")),
-    "menu.life_cycle_environments": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_environments']")),
-    "menu.red_hat_subscriptions": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_red_hat_subscriptions']")),
-    "menu.activation_keys": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_activation_keys']")),
-    "menu.red_hat_repositories": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_redhat_provider']")),
-    "menu.products": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_products']")),
-    "menu.gpg_keys": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_gpg_keys']")),
-    "menu.sync_status": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_sync_status']")),
-    "menu.sync_plans": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_sync_plans']")),
-    "menu.content_views": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_content_views']")),
-    "menu.errata": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_errata']")),
-    "menu.packages": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_packages']")),
-    "menu.puppet_modules": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_puppet_modules']")),
-    "menu.docker_tags": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_docker_tags']")),
-
-    # Containers Menu
-    "menu.containers": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='containers_menu']")),
-    "menu.all_containers": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_containers']")),
-    "menu.new_container": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_new_container']")),
-    "menu.registries": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_registries']")),
-
-    # Hosts Menu
-    "menu.hosts": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='hosts_menu']")),
-    "menu.all_hosts": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_hosts']")),
-    "menu.discovered_hosts": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_discovered_hosts']")),
-    "menu.content_hosts": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_content_hosts']")),
-    "menu.host_collections": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_host_collections']")),
-    "menu.operating_systems": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='menu_item_operatingsystems']")),
-    "menu.provisioning_templates": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_provisioning_templates']")),
-    "menu.partition_tables": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_partition_tables']")),
-    "menu.job_templates": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_job_templates']")),
-    "menu.installation_media": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_media']")),
-    "menu.hardware_models": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_models']")),
-    "menu.architectures": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_architectures']")),
-    "menu.oscap_policy": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_compliance_policies']")),
-    "menu.oscap_content": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_compliance_contents']")),
-    "menu.oscap_reports": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_compliance_reports']")),
-
-    # Configure Menu
-    "menu.configure": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='configure_menu']")),
-    "menu.host_groups": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_hostgroups']")),
-    "menu.discovery_rules": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_discovery_rules']")),
-    "menu.global_parameters": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_common_parameters']")),
-    "menu.environments": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//li[contains(@class,'menu_tab_environments')]"
-         "/a[@id='menu_item_environments']")),
-    "menu.puppet_classes": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_puppetclasses']")),
-    "menu.smart_variables": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_lookup_keys']")),
-    "menu.configure_groups": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_config_groups']")),
-
-    # Infrastructure Menu
-    "menu.infrastructure": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='infrastructure_menu']")),
-    "menu.smart_proxies": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_smart_proxies']")),
-    "menu.compute_resources": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_compute_resources']")),
-    "menu.compute_profiles": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_compute_profiles']")),
-    "menu.subnets": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_subnets']")),
-    "menu.domains": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_domains']")),
-
-    # Access Insights menu
-    "menu.insights": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@id='redhat_access_top_menu']")),
-    "insights.overview": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/redhat_access/insights']")),
-    "insights.rules": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/redhat_access/insights/rules/']")),
-    "insights.systems": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/redhat_access/insights/systems/']")),
-    "insights.manage": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/redhat_access/insights/manage']")),
-
-
-    # Administer Menu
-    "menu.administer": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='administer_menu']")),
-    "menu.ldap_auth": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_auth_source_ldaps']")),
-    "menu.users": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_users']")),
-    "menu.user_groups": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_usergroups']")),
-    "menu.roles": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_roles']")),
-    "menu.bookmarks": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_bookmarks']")),
-    "menu.settings": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_settings']")),
-    "menu.about": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//a[@id='menu_item_about_index']")),
-
-    # Account Menu
-    "menu.account": (By.XPATH, "//a[@id='account_menu']"),
-    "menu.sign_out": (By.XPATH, "//a[@id='menu_item_logout']"),
-    "menu.my_account": (By.XPATH, "//a[@id='menu_item_my_account']"),
-
-    # Common Locators for Orgs and Locations
-    "menu.any_context": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//li[contains(@class,'org-switcher')]/a")),
-    # Updated to current_text as the fetched text can also be org+loc
-    "menu.current_text": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//li[contains(@class,'org-switcher')]/a")),
-    "menu.fetch_org": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//li[contains(@class, 'org-menu')]/a")),
-    "menu.fetch_loc": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style, 'fixed')]"
-         "//li[contains(@class, 'loc-menu')]/a")),
-
-    # Orgs
-    "org.manage_org": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@class='manage-menu' and contains(@href, 'organizations')]")),
-    "org.nav_current_org": (
-        By.XPATH,
-        ("(//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//li[contains(@class,'org-switcher')]"
-         "//li/a[@data-toggle='dropdown'])[1]")),
-    "org.select_org": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/organizations/clear']/../../li/a[contains(.,'%s')]|"
-         "//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/organizations/clear']/../../li/a"
-         "/span[contains(@data-original-title, '%s')]")),
-
-    # Locations
-    "loc.manage_loc": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@class='manage-menu' and contains(@href, 'locations')]")),
-    "loc.nav_current_loc": (
-        By.XPATH,
-        ("(//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//li[contains(@class,'org-switcher')]"
-         "//li/a[@data-toggle='dropdown'])[2]")),
-    "loc.select_loc": (
-        By.XPATH,
-        ("//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/locations/clear']/../../li/a[contains(.,'%s')]|"
-         "//div[contains(@style,'static') or contains(@style,'fixed')]"
-         "//a[@href='/locations/clear']/../../li/a"
-         "/span[contains(@data-original-title, '%s')]"))
-})
-
-tab_locators = LocatorDict({
-
-    # common
-    "tab_primary": (By.XPATH, "//a[@href='#primary']"),
-    # Third level UI
-    "tab_loc": (By.XPATH, "//a[@href='#locations']"),
-    "tab_org": (By.XPATH, "//a[@href='#organizations']"),
-
-    # Operating System
-    # Third level UI
-    "operatingsys.tab_primary": (By.XPATH, "//a[@href='#primary']"),
-    "operatingsys.tab_ptable": (By.XPATH, "//a[@href='#ptable']"),
-    "operatingsys.tab_medium": (By.XPATH, "//a[@href='#media']"),
-    "operatingsys.tab_templates": (By.XPATH, "//a[@href='#templates']"),
-    "operatingsys.tab_parameters": (By.XPATH, "//a[@href='#params']"),
-
-    # Host
-    # Third level UI
-
-    "host.tab_host": (By.XPATH, "//a[@href='#primary']"),
-    "host.tab_puppet_classes": (By.XPATH, "//a[@href='#puppet_klasses']"),
-    "host.tab_interfaces": (By.XPATH, "//a[@href='#network']"),
-    "host.tab_operating_system": (
-        By.XPATH, "//form[@id='new_host']//a[@href='#os']"),
-    "host.tab_virtual_machine": (
-        By.XPATH, "//form[@id='new_host']//a[@href='#compute_resource']"),
-    "host.tab_params": (By.XPATH, "//a[@href='#params']"),
-    "host.tab_additional_information": (By.XPATH, "//a[@href='#info']"),
-
-    # Provisioning Templates
-    # Third level UI
-
-    "provision.tab_type": (By.XPATH, "//a[@href='#template_type']"),
-    "provision.tab_association": (By.XPATH,
-                                  "//a[@href='#template_associations']"),
-    "provision.tab_history": (By.XPATH, "//a[@href='#history']"),
-
-    # Job Templates
-    # Third level UI
-
-    "job.tab_job": (By.XPATH, "//a[@href='#template_job']"),
-    "job.tab_type": (By.XPATH, "//a[@href='#template_type']"),
-    "job.tab_history": (By.XPATH, "//a[@href='#history']"),
-
-    # Users
-    # Third level UI
-
-    "users.tab_primary": (By.XPATH, "//a[@href='#primary']"),
-    "users.tab_roles": (By.XPATH, "//a[@href='#roles']"),
-    "users.tab_locations": (By.XPATH, "//a[@href='#locations']"),
-    "users.tab_organizations": (By.XPATH, "//a[@href='#organizations']"),
-    "users.tab_filters": (By.XPATH, "//a[@href='#filters']"),
-
-    "prd.tab_details": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'info')]"),
-    "prd.tab_repos": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'repositories')]"),
-    "prd.tab_tasks": (
-        By.XPATH, ("//a[@class='ng-scope' and contains(@href,'tasks')"
-                   " and contains(@ui-sref, 'products')]")),
-
-    # For Orgs and Locations
-    "context.tab_users": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'users')]"),
-    "context.tab_capsules": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'smart_proxies')]"),
-    "context.tab_subnets": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'subnets')]"),
-    "context.tab_resources": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'resources')]"),
-    "context.tab_media": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'media')]"),
-    "context.tab_template": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'template')]"),
-    "context.tab_ptable": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'ptables')]"),
-    "context.tab_domains": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'domains')]"),
-    "context.tab_env": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'environments')]"),
-    "context.tab_hostgrps": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'hostgroups')]"),
-    "context.tab_locations": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'locations')]"),
-    "context.tab_organizations": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'organizations')]"),
-    "context.tab_parameters": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href,'params')]"),
-
-    # Roles
-    # Third level UI
-    "roles.tab_filter": (
-        By.XPATH, "//a[@href='#primary']"),
-    "roles.tab_org": (
-        By.XPATH, "//a[@href='#organizations']"),
-
-    # GPG key
-    # Third level UI
-    "gpgkey.tab_details": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'info')]"),
-    "gpgkey.tab_products": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'products')]"),
-    "gpgkey.tab_repos": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'repositories')]"),
-
-    # Content Views
-    # Third level UI
-    "contentviews.tab_details": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'info')]"),
-    "contentviews.tab_versions": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@ui-sref, 'details.versions')]"),
-    "contentviews.tab_content": (
-        By.XPATH, "//ul/li/a[@class='dropdown-toggle']/i"),
-    "contentviews.tab_content_views": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@href, 'content-views')]"),
-    "contentviews.tab_puppet_modules": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@href, 'puppet_modules')]"),
-    "contentviews.tab_docker_content": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@href, 'docker')]"),
-    "contentviews.tab_history": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href, 'history')]"),
-    "contentviews.tab_tasks": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href, 'tasks')]"),
-    "contentviews.tab_repo_add": (
-        By.XPATH,
-        "//a[contains(@href, 'repositories') and "
-        "contains(@ui-sref, 'available')]"),
-    "contentviews.tab_repo_remove": (
-        By.XPATH,
-        ("//a[contains(@href, 'repositories') and contains(@ui-sref, 'list')]"
-         "/span[@class='ng-scope' and contains(., 'List/Remove')]")),
-    "contentviews.tab_cv_add": (
-        By.XPATH,
-        "//a[contains(@ui-sref, 'content-views.available')]"),
-    "contentviews.tab_cv_remove": (
-        By.XPATH, "//a[contains(@ui-sref, 'content-views.list')]"),
-    "contentviews.tab_pkg_group_add": (
-        By.XPATH, "//a[contains(@ui-sref, 'package_group.available')]"),
-    "contentviews.tab_pkg_group_remove": (
-        By.XPATH, "//a[contains(@ui-sref, 'package_group.list')]"),
-    "contentviews.tab_add": (
-        By.XPATH, "//a[contains(@ui-sref, 'available')]"),
-    "contentviews.tab_remove": (
-        By.XPATH, "//a[contains(@ui-sref, 'list')]"),
-
-    # Content Hosts
-    # Third level UI
-    "contenthost.tab_details": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'info')]"),
-    "contenthost.tab_subscriptions": (
-        By.XPATH,
-        ("//li[@class='dropdown' and contains(@ng-class, "
-         "'content-hosts.details.subscriptions.list')]/a")),
-    "contenthost.tab_subscriptions_subscriptions": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@href,'subscriptions')]"),
-    "contenthost.tab_packages": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@href,'packages')]"),
-    "contenthost.tab_errata": (
-        By.XPATH,
-        "//a[@class='ng-scope' and contains(@href,'errata')]"),
-    # Fourth level UI
-    "contenthost.list_subscriptions": (
-        By.XPATH,
-        ("//a[contains(@ui-sref,'subscriptions.list')]"
-         "/span[@class='ng-scope']")),
-    "contenthost.add_subscription": (
-        By.XPATH,
-        ("//a[contains(@ui-sref,'subscriptions.add')]"
-         "/span[@class='ng-scope']")),
-
-    # Sync Plans
-    # Third level UI
-    "sp.tab_details": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'info')]"),
-    "sp.tab_products": (
-        By.XPATH, "//a[@class='ng-scope' and contains(@href,'products')]"),
-    # Fourth level UI
-    "sp.list_prd": (
-        By.XPATH, "//a[contains(@ui-sref,'list')]/span[@class='ng-scope']"),
-    "sp.add_prd": (
-        By.XPATH, "//a[contains(@ui-sref,'add')]/span[@class='ng-scope']"),
-
-    # Activation Keys
-    # Third level UI
-    "ak.details": (
-        By.XPATH, "//a[contains(@href, 'info')]"),
-    "ak.subscriptions": (
-        By.XPATH, "//a[contains(@href, 'subscriptions')]/span"),
-    "ak.subscriptions_add": (
-        By.XPATH, "//a[contains(@ui-sref, 'subscriptions.add')]"),
-    "ak.subscriptions_remove": (
-        By.XPATH, "//a[contains(@ui-sref, 'subscriptions.list')]"),
-    "ak.host_collections": (
-        By.XPATH, "//a[contains(@href, 'host-collections')]"),
-    "ak.host_collections.add": (
-        By.XPATH, "//a[contains(@ui-sref, 'host-collections.add')]"),
-    "ak.host_collections.add.select": (
-        By.XPATH,
-        "//tr/td/a[contains(., '%s')]"
-        "/parent::*/parent::*"
-        "/td/input[@ng-model='hostCollection.selected']"),
-    "ak.host_collections.add.add_selected": (
-        By.XPATH, "//button[@ng-click='addHostCollections()']"),
-    "ak.host_collections.list": (
-        By.XPATH, "//a[contains(@ui-sref, 'host-collections.list')]"),
-    "ak.associations": (
-        By.XPATH, "//ul/li[@class='dropdown']/a"),
-    "ak.tab_prd_content": (
-        By.XPATH, "//a[contains(@ui-sref, 'details.products')]/span/span"),
-
-    # Manifest / subscriptions
-    "subs.tab_details": (
-        By.XPATH, "//a[contains(@ui-sref,'manifest.details')]"),
-    "subs.import_history": (
-        By.XPATH, "//a[contains(@ui-sref,'manifest.history')]"),
-    "subs.import_history.imported.success": (
-        By.XPATH, "//td[text()[contains(.,'imported successfully')]]"),
-    "subs.import_history.deleted": (
-        By.XPATH, "//td[text()[contains(., 'deleted')]]"),
-
-    # Oscap Policy
-    "oscap.content": (
-        By.XPATH, "//a[@href='#scap_content']"),
-    "oscap.schedule": (
-        By.XPATH, "//a[contains(@href, 'scap_schedule')]"),
-
-    # Subnet
-    "subnet.tab_domain": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'domains')]"),
-
-    # Settings
-    "settings.tab_general": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'General')]"),
-    "settings.tab_auth": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'Auth')]"),
-    "settings.tab_bootdisk": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'Bootdisk')]"),
-    "settings.tab_puppet": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'Puppet')]"),
-    "settings.tab_discovered": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'Discovered')]"),
-    "settings.tab_foremantasks": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href, 'ForemanTasks')]"),
-    "settings.tab_provisioning": (
-        By.XPATH,
-        "//a[@data-toggle='tab' and contains(@href, 'Provisioning')]"),
-    "puppetclass.parameters": (
-        By.XPATH, "//a[contains(@href,'class_param')]"),
-
-    # LDAP Authentication
-    "ldapserver.tab_account": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'account')]"),
-    "ldapserver.tab_attributes": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'attributes')]"),
-
-    # UserGroups
-    "usergroups.tab_roles": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'roles')]"),
-    "usergroups.tab_external": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'external')]"),
-
-    # Host Collections
-    "hostcollection.details": (
-        By.XPATH, "//a[contains(@href, 'info')]/span"),
-    "hostcollection.hosts": (
-        By.XPATH, "//a[contains(@href, 'hosts')]/span"),
-    "hostcollection.tab_host_add": (
-        By.XPATH,
-        "//a[contains(@href, 'add-hosts') and "
-        "contains(@ui-sref, 'host-collections.details.hosts.add')]"),
-    "hostcollection.tab_host_remove": (
-        By.XPATH,
-        "//a[contains(@href, 'hosts') and "
-        "contains(@ui-sref, 'host-collections.details.hosts.list')]"),
-    "hostcollection.collection_actions": (
-        By.XPATH, "//a[contains(@href, 'actions')]/span"),
-
-    # Compute resources
-    "resource.tab_containers": (
-        By.XPATH, "//a[@data-toggle='tab' and contains(@href, 'vms')]"),
-})
-
-common_locators = LocatorDict({
-
-    # common locators
-
-    # Notifications
-    "notif.error": (
-        By.XPATH, "//div[contains(@class, 'jnotify-notification-error')]"),
-    "notif.warning": (
-        By.XPATH, "//div[contains(@class, 'jnotify-notification-warning')]"),
-    "notif.success": (
-        By.XPATH, "//div[contains(@class, 'jnotify-notification-success')]"),
-    "notif.close": (
-        By.XPATH, "//a[@class='jnotify-close']"),
-
-    "alert.success": (
-        By.XPATH, "//div[contains(@class, 'alert-success')]"),
-    "alert.error": (
-        By.XPATH, "//div[contains(@class, 'alert-danger')]"),
-    "alert.success_sub_form": (
-        By.XPATH,
-        "//div[@ng-hide]//div[contains(@class, 'alert-success')]"),
-    "alert.error_sub_form": (
-        By.XPATH,
-        "//div[@ng-hide]//div[contains(@class, 'alert-danger')]"),
-
-    "selected_entity": (
-        By.XPATH,
-        ("//div[@class='ms-selection']/ul[@class='ms-list']"
-         "/li[@class='ms-elem-selection ms-selected']")),
-    "select_filtered_entity": (
-        By.XPATH, "//table//a/span[contains(@data-original-title, '%s')]"),
-    "checked_entity": (
-        By.XPATH, "//input[@checked='checked']/parent::label"),
-    "entity_select": (
-        By.XPATH,
-        ("//div[@class='ms-selectable']//"
-         "li[not(contains(@style, 'display: none'))]/span[contains(.,'%s')]")),
-    "entity_deselect": (
-        By.XPATH,
-        ("//div[@class='ms-selection']//"
-         "li[not(contains(@style, 'display: none'))]/span[contains(.,'%s')]")),
-    "entity_checkbox": (
-        By.XPATH,
-        "//label[normalize-space(.)='%s']/input[@type='checkbox']"),
-    "entity_select_list": (
-        By.XPATH,
-        "//ul/li/div[normalize-space(.)='%s']"),
-    "select_list_search_box": (
-        By.XPATH, "//div[@id='select2-drop']//input"),
-    "name_haserror": (
-        By.XPATH,
-        ("//label[@for='name']/../../"
-         "div[contains(@class,'has-error')]")),
-    "haserror": (
-        By.XPATH,
-        "//div[contains(@class,'has-error')]"),
-    "common_haserror": (
-        By.XPATH,
-        ("//span[@class='help-block']/ul/"
-         "li[contains(@ng-repeat,'error.messages')]")),
-    "common_invalid": (
-        By.XPATH,
-        "//input[@id='name' and contains(@class,'ng-invalid')]"),
-    "common_param_error": (
-        By.XPATH,
-        ("//div[@id='parameters']/span[@class='help-block'"
-         "and string-length(text()) > 10]")),
-
-    "search": (By.ID, "search"),
-    "auto_search": (
-        By.XPATH, "//ul[contains(@id, 'ui-id')]/li/a[contains(., '%s')]"),
-    "search_button": (By.XPATH, "//button[contains(@type,'submit')]"),
-    "search_dropdown": (
-        By.XPATH,
-        ("//button[contains(@class, 'dropdown-toggle')]"
-         "[@data-toggle='dropdown']")),
-    "cancel_form": (By.XPATH, "//a[text()='Cancel']"),
-    "submit": (By.NAME, "commit"),
-    "filter": (By.XPATH,
-               ("//div[@id='ms-%s_ids']"
-                "//input[contains(@class,'ms-filter')]")),
-    "parameter_tab": (By.XPATH, "//a[contains(., 'Parameters')]"),
-    "add_parameter": (
-        By.XPATH, "//a[contains(text(),'+ Add Parameter')]"),
-    "parameter_name": (
-        By.XPATH, "//input[@placeholder='Name' and not(@value)]"),
-    "parameter_value": (
-        By.XPATH, "//textarea[@placeholder='Value' and not(text())]"),
-    "parameter_remove": (
-        By.XPATH, "//tr/td/input[@value='%s']/following::td/a"),
-
-    # Katello Common Locators
-    "confirm_remove": (
-        By.XPATH, "//button[@ng-click='ok()' or @ng-click='delete()']"),
-    "create": (By.XPATH, "//button[contains(@ng-click,'Save')]"),
-    "save": (
-        By.XPATH, ("//button[contains(@ng-click,'save')"
-                   "and not(contains(@class,'ng-hide'))]")),
-    "cancel": (By.XPATH, "//button[@aria-label='Close']"),
-    "name": (By.ID, "name"),
-    "label": (By.ID, "label"),
-    "description": (By.ID, "description"),
-    "kt_search": (By.XPATH, "//input[@ng-model='table.searchTerm']"),
-    "kt_search_button": (
-        By.XPATH,
-        "//button[@ng-click='table.search(table.searchTerm)']"),
-    # Katello common Product and Repo locators
-    "gpg_key": (By.ID, "gpg_key_id"),
-    "all_values": (By.XPATH,
-                   ("//div[contains(@class,'active')]//input[@type='checkbox'"
-                    " and contains(@name, '%s')]")),
-    "all_values_selection": (
-        By.XPATH,
-        ("//div[@class='ms-selection']//ul[@class='ms-list']/li"
-         "/span[contains(.,'%s')]/..")),
-    "usage_limit": (
-        By.XPATH,
-        "//input[contains(@ng-model, 'max')"
-        "and contains(@ng-model, 'hosts')]"),
-    "usage_limit_checkbox": (
-        By.XPATH,
-        "//input[contains(@ng-model, 'unlimited')"
-        "and contains(@ng-model, 'hosts')]"),
-    "invalid_limit": (
-        By.XPATH,
-        "//input[contains(@id, 'max') and contains(@class, 'ng-invalid')]"),
-    "modal_background": (
-        By.XPATH,
-        "//*[@class='modal-backdrop fade in']",
-    ),
-    "select_repo": (By.XPATH, "//select[@ng-model='repository']"),
-
-    # ace editor
-    "ace.input": (By.XPATH, "//label[contains(., 'Input') and"
-                            " contains(@class, 'btn')]"),
-    "ace.diff": (By.XPATH, "//label[contains(., 'Diff') and"
-                           " contains(@class, 'btn')]"),
-    "ace.preview": (By.XPATH, "//label[contains(., 'Preview') and"
-                              " contains(@class, 'btn')]"),
-
-    # 'Run Job' button that is accessible from Jobs and Hosts pages
-    "run_job": (By.XPATH, "//a[@data-id='aid_job_invocations_new']"),
-})
 
 locators = LocatorDict({
